+++ Domain/Common/AuditableBaseEntity.cs
@@ -4,10 +4,10 @@ namespace Domain.Common
 {
     public class AuditableBaseEntity
     {
-        public virtual long Id { get; set; }
-        public virtual long CreatedBy { get; set; }
+        public virtual int Id { get; set; }
+        public virtual int CreatedBy { get; set; }
         public virtual DateTime CreatedAt { get; set; }
-        public virtual long UpdatedBy { get; set; }
+        public virtual int UpdatedBy { get; set; }
         public virtual DateTime UpdatedAt { get; set; }
 
     }
